# runtime/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AMS API",
    "version": "0.2.0"
  },
  "paths": {
    "/api/platform/auth/sign-in": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Platform Sign In",
        "operationId": "platform_sign_in_api_platform_auth_sign_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformSignInInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/demo": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Demo",
        "operationId": "demo_api_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password/sign-in": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Password Sign In",
        "operationId": "password_sign_in_api_auth_password_sign_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordSignInInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/password/change-required": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Password Change",
        "operationId": "password_change_api_auth_password_change_required_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invitations/accept": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Accept",
        "operationId": "accept_api_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/academies": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Academies",
        "operationId": "academies_api_platform_academies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AcademyDto"
                  },
                  "type": "array",
                  "title": "Response Academies Api Platform Academies Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Create Academy",
        "operationId": "create_academy_api_platform_academies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademyCreatedDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/platform/credential-handoffs/{handoff_id}/reveal": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Reveal Credentials",
        "operationId": "reveal_credentials_api_platform_credential_handoffs__handoff_id__reveal_post",
        "parameters": [
          {
            "name": "handoff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Handoff Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffSecretDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/credential-handoffs/{handoff_id}/copied": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Copied Credentials",
        "operationId": "copied_credentials_api_platform_credential_handoffs__handoff_id__copied_post",
        "parameters": [
          {
            "name": "handoff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Handoff Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/platform/credential-handoffs/{handoff_id}/retry-email": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Retry Credentials Email",
        "operationId": "retry_credentials_email_api_platform_credential_handoffs__handoff_id__retry_email_post",
        "parameters": [
          {
            "name": "handoff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Handoff Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DemoInput": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "userId"
        ],
        "title": "DemoInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChangeInput": {
        "properties": {
          "newPassword": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Newpassword"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "newPassword"
        ],
        "title": "PasswordChangeInput"
      },
      "PasswordSignInInput": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "PasswordSignInInput"
      },
...
      "PlatformSignInInput": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "PlatformSignInInput"
      },
...
      "TokenDto": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          }
        },
        "type": "object",
        "required": [
          "accessToken"
        ],
        "title": "TokenDto"
      },
...
    }
  }
}